moved filters and search Params to parent level
