+++ frontend/micro-ui/web/micro-ui-internals/packages/libraries/src/utils/pt/index.js
@@ -63,3 +63,16 @@ export const convertDateToEpoch = (dateString, dayStartOrEnd = "dayend") => {
     return dateString;
   }
 };
+
+export const convertEpochToDate = (dateEpoch) => {
+  if (dateEpoch == null || dateEpoch == undefined || dateEpoch == "") {
+      return "NA";
+  }
+  const dateFromApi = new Date(dateEpoch);
+  let month = dateFromApi.getMonth() + 1;
+  let day = dateFromApi.getDate();
+  let year = dateFromApi.getFullYear();
+  month = (month > 9 ? "" : "0") + month;
+  day = (day > 9 ? "" : "0") + day;
+  return `${day}/${month}/${year}`;
+};
