+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/Estimate/src/pageComponents/ViewProject.js
@@ -3,28 +3,12 @@ import React, { Fragment, useState, useEffect } from 'react'
 import { useTranslation } from 'react-i18next';
 import ApplicationDetails from '../../../templates/ApplicationDetails';
 
-const ViewProject = (props) => {
+const ViewProject = ({searchParams, filters}) => {
   const tenantId =  Digit.ULBService.getCurrentTenantId();
-  const queryStrings = Digit.Hooks.useQueryParams();
-  const headerLocale = Digit.Utils.locale.getTransformedLocale(tenantId);
-
-  const searchParams = {
-    Projects : [
-        {   //will update this once integration for view is completed
-            tenantId : queryStrings?.tenantId,
-            projectNumber : queryStrings?.projectNumber
-        }
-    ]
-  } 
-  const filters = {
-      limit : 10,
-      offset : 0,
-      includeAncestors : true,
-      includeDescendants : true
-  }
-    
+  const headerLocale = Digit.Utils.locale.getTransformedLocale(tenantId);    
   const { t } = useTranslation()
   const { data, isLoading } = Digit.Hooks.works.useViewProjectDetailsInEstimate(t, tenantId, searchParams, filters, headerLocale);
+
   return (
     <>
         <ApplicationDetails
