+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/Project/src/pageComponents/ViewFinancialDetails.js
@@ -3,28 +3,12 @@ import React, { Fragment, useState, useEffect } from 'react'
 import { useTranslation } from 'react-i18next';
 import ApplicationDetails from '../../../templates/ApplicationDetails';
 
-const ViewFinancialDetails = (props) => {
+const ViewFinancialDetails = ({searchParams, filters}) => {
   const tenantId =  Digit.ULBService.getCurrentTenantId();
-  const queryStrings = Digit.Hooks.useQueryParams();
   const headerLocale = Digit.Utils.locale.getTransformedLocale(tenantId);
-
-  const searchParams = {
-    Projects : [
-        {
-            tenantId : queryStrings?.tenantId,
-            projectNumber : queryStrings?.projectNumber
-        }
-    ]
-  } 
-  const filters = {
-      limit : 10,
-      offset : 0,
-      includeAncestors : true,
-      includeDescendants : true
-  }
-    
   const { t } = useTranslation()
   const { data, isLoading } = Digit.Hooks.works.useViewProjectDetailsInEstimate(t, tenantId, searchParams, filters, headerLocale);
+  
   return (
     <>
         <ApplicationDetails
