+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/Project/src/pageComponents/ViewSubProjectsDetails.js
@@ -1,19 +1,7 @@
 import React, { Fragment, useState } from "react";
 import { useTranslation } from "react-i18next";
+import { convertEpochToDate } from "../../../../libraries/src/utils/pt";
 import PropertyDocuments from "../../../templates/ApplicationDetails/components/PropertyDocuments";
-
-const convertEpochToDate = (dateEpoch) => {
-    if (dateEpoch == null || dateEpoch == undefined || dateEpoch == "") {
-        return "NA";
-    }
-    const dateFromApi = new Date(dateEpoch);
-    let month = dateFromApi.getMonth() + 1;
-    let day = dateFromApi.getDate();
-    let year = dateFromApi.getFullYear();
-    month = (month > 9 ? "" : "0") + month;
-    day = (day > 9 ? "" : "0") + day;
-    return `${day}/${month}/${year}`;
-};
 
 const ViewSubProjectsDetails = ({subProjects}) => {
     const {t} = useTranslation();
