+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/Project/src/pages/employee/ProjectDetails/index.js
@@ -88,6 +88,8 @@ const ProjectDetails = () => {
               <ProjectDetailsNavDetails 
                 activeLink={activeLink}
                 subProjects={subProjects}
+                searchParams={searchParams}
+                filters={filters}
               />
             </HorizontalNav>
             <ActionBar>
